+++ gmall-seckill/src/main/java/com/hhx4/gmall/seckill/vo/SeckillSkuVo.java
@@ -11,31 +11,32 @@ import java.math.BigDecimal;
 public class SeckillSkuVo {
     private Long id;
     /**
-     *
+     * 活动id
      */
     private Long promotionId;
+
     /**
-     *
+     * 活动场次id
      */
     private Long promotionSessionId;
     /**
-     *
+     * 商品id
      */
     private Long skuId;
     /**
-     *
+     * 秒杀价格
      */
     private BigDecimal seckillPrice;
     /**
-     *
+     * 秒杀总量
      */
-    private BigDecimal seckillCount;
+    private Integer seckillCount;
     /**
-     * ÿ
+     * 每人限购数量
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
-     *
+     * 排序
      */
     private Integer seckillSort;
 }
